[RED] Merge analyze test to respective test
