--- backend/tests/services/analyzeInvoiceService.test.js
@@ -1,115 +1,0 @@
-const { analyzeInvoice } = require("../../src/services/analyzeInvoiceService");
-const { DocumentAnalysisClient } = require("@azure/ai-form-recognizer");
-
-jest.mock("@azure/ai-form-recognizer");
-
-describe("Invoice Analysis Service", () => {
-  const documentUrl = "https://slicedinvoices.com/pdf/wordpress-pdf-invoice-plugin-sample.pdf";
-
-  beforeEach(() => {
-    jest.clearAllMocks();
-  });
-
-  describe("Positive Cases", () => {
-    test("should successfully process a valid PDF", async () => {
-      const mockClient = {
-        beginAnalyzeDocument: jest.fn().mockResolvedValue({
-          pollUntilDone: jest.fn().mockResolvedValue({ success: true }),
-        }),
-      };
-
-      DocumentAnalysisClient.mockImplementation(() => mockClient);
-
-      const result = await analyzeInvoice(documentUrl);
-      expect(result).toEqual({
-        message: "PDF processed successfully",
-        data: { success: true },
-      });
-    });
-  });
-
-  describe("Negative Cases", () => {
-    test("should throw an error if documentUrl is missing", async () => {
-      await expect(analyzeInvoice()).rejects.toThrow("documentUrl is required");
-    });
-
-    test("should throw 'Service is temporarily unavailable' if API returns 503", async () => {
-      const mockError = new Error("Service Unavailable");
-      mockError.statusCode = 503;
-
-      const mockClient = {
-        beginAnalyzeDocument: jest.fn().mockRejectedValue(mockError),
-      };
-
-      DocumentAnalysisClient.mockImplementation(() => mockClient);
-
-      await expect(analyzeInvoice(documentUrl)).rejects.toThrow(
-        "Service is temporarily unavailable. Please try again later."
-      );
-    });
-
-    test("should throw 'Conflict error occurred' if API returns 409", async () => {
-      const mockError = new Error("Conflict Error");
-      mockError.statusCode = 409;
-
-      const mockClient = {
-        beginAnalyzeDocument: jest.fn().mockRejectedValue(mockError),
-      };
-
-      DocumentAnalysisClient.mockImplementation(() => mockClient);
-
-      await expect(analyzeInvoice(documentUrl)).rejects.toThrow(
-        "Conflict error occurred. Please check the document and try again."
-      );
-    });
-
-    test("should throw 'Failed to process the document' for any other API error", async () => {
-      const mockError = new Error("Unknown Error");
-      mockError.statusCode = 500;
-
-      const mockClient = {
-        beginAnalyzeDocument: jest.fn().mockRejectedValue(mockError),
-      };
-
-      DocumentAnalysisClient.mockImplementation(() => mockClient);
-
-      await expect(analyzeInvoice(documentUrl)).rejects.toThrow(
-        "Failed to process the document"
-      );
-    });
-  });
-
-  describe("Corner Cases", () => {
-    test("should handle empty response from API gracefully", async () => {
-      const mockClient = {
-        beginAnalyzeDocument: jest.fn().mockResolvedValue({
-          pollUntilDone: jest.fn().mockResolvedValue(null),
-        }),
-      };
-
-      DocumentAnalysisClient.mockImplementation(() => mockClient);
-
-      const result = await analyzeInvoice(documentUrl);
-      expect(result).toEqual({
-        message: "PDF processed successfully",
-        data: null,
-      });
-    });
-
-    test("should handle API response with missing fields", async () => {
-      const mockClient = {
-        beginAnalyzeDocument: jest.fn().mockResolvedValue({
-          pollUntilDone: jest.fn().mockResolvedValue({}),
-        }),
-      };
-
-      DocumentAnalysisClient.mockImplementation(() => mockClient);
-
-      const result = await analyzeInvoice(documentUrl);
-      expect(result).toEqual({
-        message: "PDF processed successfully",
-        data: {},
-      });
-    });
-  });
-});
